+++ jhj-utils/src/main/java/com/meijia/utils/poi/POIUtils.java
@@ -85,7 +85,7 @@ public class POIUtils {
 	        for (short i = 1; i < list.size(); i++) {
 	            // Row 行,Cell 方格 , Row 和 Cell 都是从0开始计数的
 	            // 创建一行，在页sheet上
-	            Row row1 = sheet.createRow((short) i);
+	            Row row1 = sheet.createRow(i);
 	            // 在row行上创建一个方格
 	            for(short j=0;j<keys.length;j++){
 	                Cell cell = row1.createCell(j);
